+++ tests/AlertTest.php
@@ -52,4 +52,36 @@ Message1
 HTML;
         $this->assertEqualsWithoutLE($expectedHtml, $html);
     }
+
+    /**
+     * @see https://github.com/yiisoft/yii2-bootstrap5/issues/11
+     */
+    public function testDismissibleAlertCustomButton()
+    {
+        Alert::$counter = 0;
+        $html = Alert::widget([
+            'body' => "Low Blow: Bob Loblaw's Law Blog Lobs Law Bomb",
+            'options' => ['class' => 'alert-warning'],
+            'closeButton' => [
+                'label' => 'Dismiss',
+                'tag' => 'a',
+                'class' => ['widget' => 'btn btn-outline-warning'],
+                'style' => [
+                    'position' => 'absolute',
+                    'top' => '.5rem',
+                    'right' => '.5rem'
+                ],
+            ]
+        ]);
+
+        $expectedHtml = <<<HTML
+<div id="w0" class="alert-warning alert alert-dismissible" role="alert">
+
+Low Blow: Bob Loblaw's Law Blog Lobs Law Bomb
+<a class="btn btn-outline-warning" data-bs-dismiss="alert" aria-label="Close" style="position: absolute; top: .5rem; right: .5rem;">Dismiss</a>
+
+</div>
+HTML;
+        $this->assertEqualsWithoutLE($expectedHtml, $html);
+    }
 }