+++ tests/ToastTest.php
@@ -23,7 +23,7 @@ class ToastTest extends TestCase
 <div id="w0" class="toast" role="alert" aria-live="assertive" aria-atomic="true">
 <div class="toast-header">
 <strong class="me-auto"></strong>
-<button type="button" class="btn-close" aria-label="Close" data-bs-dismiss="toast"></button>
+<button type="button" class="btn-close" data-bs-dismiss="toast" aria-label="Close"></button>
 </div>
 <div class="toast-body test" style="text-align: center;">
 
@@ -55,7 +55,7 @@ HTML;
 <div class="toast-header">
 <strong class="me-auto">Toast title</strong>
 <small class="text-muted">a minute ago</small>
-<button type="button" class="btn-close" aria-label="Close" data-bs-dismiss="toast"></button>
+<button type="button" class="btn-close" data-bs-dismiss="toast" aria-label="Close"></button>
 </div>
 <div class="toast-body">
 Woohoo, you're reading this text in a toast!
@@ -81,7 +81,7 @@ HTML;
 <div class="toast-header">
 <strong class="me-auto">Toast title</strong>
 <small class="toast-date-time text-muted" style="text-align: right;">a minute ago</small>
-<button type="button" class="btn-close" aria-label="Close" data-bs-dismiss="toast"></button>
+<button type="button" class="btn-close" data-bs-dismiss="toast" aria-label="Close"></button>
 </div>
 <div class="toast-body">
 
@@ -105,7 +105,7 @@ HTML;
 <div id="w0" class="toast" role="alert" aria-live="assertive" aria-atomic="true">
 <div class="toast-header">
 <h5 class="me-auto" style="text-align: left;">Toast title</h5>
-<button type="button" class="btn-close" aria-label="Close" data-bs-dismiss="toast"></button>
+<button type="button" class="btn-close" data-bs-dismiss="toast" aria-label="Close"></button>
 </div>
 <div class="toast-body">