+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Yii Framework 2 bootstrap5 extension Change Log
 -----------------------
 
 - Bug #5: BootstrapWidgetTrait::registerPlugin do nothing if no clientOptions is provided (dicrtarasov)
+- Enh #11: Brought back close button api (simialbi)
+- Bug #6: yii\bootstrap5\BaseHtml::staticControl(): Argument #1 ($value) must be of type string, int given (dicrtarasov)
+- Bug #9: fixed default ActiveField::hintOptions (dicrtarasov)
 
 
 2.0.1 August 11, 2021